split util functions into `utils.js` and fix hashes for snx.issued/burned
