+++ index.js
@@ -5,6 +5,8 @@ const { SubscriptionClient } = require('subscriptions-transport-ws');
 
 const pageResults = require('graph-results-pager');
 
+const { ZERO_ADDRESS, hexToAscii, roundTimestampTenSeconds, getHashFromId } = require('./utils');
+
 const graphAPIEndpoints = {
 	snx: 'https://api.thegraph.com/subgraphs/name/synthetixio-team/synthetix',
 	depot: 'https://api.thegraph.com/subgraphs/name/synthetixio-team/synthetix-depot',
@@ -21,22 +23,6 @@ const graphWSEndpoints = {
 	exchanges: 'wss://api.thegraph.com/subgraphs/name/synthetixio-team/synthetix-exchanges',
 	rates: 'wss://api.thegraph.com/subgraphs/name/synthetixio-team/synthetix-rates',
 };
-
-const ZERO_ADDRESS = '0x' + '0'.repeat(40);
-
-const hexToAscii = str => {
-	const hex = str.toString();
-	let out = '';
-	for (let n = 2; n < hex.length; n += 2) {
-		const nextPair = hex.substr(n, 2);
-		if (nextPair !== '00') {
-			out += String.fromCharCode(parseInt(nextPair, 16));
-		}
-	}
-	return out;
-};
-
-const roundTimestampTenSeconds = timestamp => Math.round(timestamp / 10) * 10;
 
 module.exports = {
 	pageResults,
@@ -61,7 +47,7 @@ module.exports = {
 			})
 				.then(results =>
 					results.map(({ id, user, amount, type, minimum, depositIndex, block, timestamp }) => ({
-						hash: id.split('-')[0],
+						hash: getHashFromId(id),
 						user,
 						amount: amount / 1e18,
 						type,
@@ -103,7 +89,7 @@ module.exports = {
 			})
 				.then(results =>
 					results.map(({ id, fromAddress, toAddress, fromETHAmount, toAmount, depositIndex, block, timestamp }) => ({
-						hash: id.split('-')[0],
+						hash: getHashFromId(id),
 						fromAddress,
 						toAddress,
 						fromETHAmount: fromETHAmount / 1e18,
@@ -112,6 +98,7 @@ module.exports = {
 						block: Number(block),
 						timestamp: Number(timestamp * 1000),
 						date: new Date(timestamp * 1000),
+						type: 'cleared',
 					})),
 				)
 				.catch(err => console.error(err));
@@ -135,7 +122,7 @@ module.exports = {
 			})
 				.then(results =>
 					results.map(({ id, from, fromAmount, fromCurrency, toAmount, toCurrency, block, timestamp }) => ({
-						hash: id.split('-')[0],
+						hash: getHashFromId(id),
 						from,
 						fromAmount: fromAmount / 1e18,
 						fromCurrency,
@@ -144,6 +131,7 @@ module.exports = {
 						block: Number(block),
 						timestamp: Number(timestamp * 1000),
 						date: new Date(timestamp * 1000),
+						type: 'bought',
 					})),
 				)
 				.catch(err => console.error(err));
@@ -185,7 +173,7 @@ module.exports = {
 			block: Number(block),
 			timestamp: Number(timestamp * 1000),
 			date: new Date(timestamp * 1000),
-			hash: id.split('-')[0],
+			hash: getHashFromId(id),
 			fromAddress: from,
 			fromAmount: fromAmount / 1e18, // shorthand way to convert wei into eth
 			fromCurrencyKeyBytes: fromCurrencyKey,
@@ -322,7 +310,7 @@ module.exports = {
 							block: Number(block),
 							timestamp: Number(timestamp * 1000),
 							date: new Date(timestamp * 1000),
-							hash: id.split('-')[0],
+							hash: getHashFromId(id),
 							account,
 							amount: amount / 1e18, // shorthand way to convert wei into eth,
 							amountInUSD: amountInUSD / 1e18,
@@ -423,7 +411,7 @@ module.exports = {
 						block: Number(block),
 						timestamp: Number(timestamp * 1000),
 						date: new Date(timestamp * 1000),
-						hash: id.split('-')[0],
+						hash: getHashFromId(id),
 						from,
 						to,
 						value: value / 1e18,
@@ -455,7 +443,7 @@ module.exports = {
 			})
 				.then(results =>
 					results.map(({ id, balanceOf, synth }) => ({
-						address: id.split('-')[0],
+						address: getHashFromId(id),
 						balanceOf: balanceOf ? balanceOf / 1e18 : null,
 						synth,
 					})),
@@ -525,7 +513,7 @@ module.exports = {
 						synth,
 						timestamp: Number(timestamp * 1000),
 						date: new Date(timestamp * 1000),
-						hash: id.split('-')[0],
+						hash: getHashFromId(id),
 						rate: rate / 1e18,
 					})),
 				)
@@ -647,11 +635,12 @@ module.exports = {
 			})
 				.then(results =>
 					results.map(({ id, account, timestamp, block, value }) => ({
-						hash: id,
+						hash: getHashFromId(id),
 						account,
 						timestamp: Number(timestamp * 1000),
 						block: Number(block),
 						value: value / 1e18,
+						type: 'issued',
 					})),
 				)
 				.catch(err => console.error(err));
@@ -682,11 +671,12 @@ module.exports = {
 			})
 				.then(results =>
 					results.map(({ id, account, timestamp, block, value }) => ({
-						hash: id,
+						hash: getHashFromId(id),
 						account,
 						timestamp: Number(timestamp * 1000),
 						block: Number(block),
 						value: value / 1e18,
+						type: 'burned',
 					})),
 				)
 				.catch(err => console.error(err));
@@ -862,7 +852,7 @@ module.exports = {
 						block: Number(block),
 						timestamp: Number(timestamp * 1000),
 						date: new Date(timestamp * 1000),
-						hash: id.split('-')[0],
+						hash: getHashFromId(id),
 						from,
 						to,
 						value: value / 1e18,
@@ -895,12 +885,13 @@ module.exports = {
 			})
 				.then(results =>
 					results.map(({ id, account, timestamp, block, value, rewards }) => ({
-						hash: id.split('-')[0],
+						hash: getHashFromId(id),
 						account,
 						timestamp: Number(timestamp * 1000),
 						block: Number(block),
 						value: value / 1e18,
 						rewards: rewards / 1e18,
+						type: 'feesClaimed',
 					})),
 				)
 				.catch(err => console.error(err));
@@ -1036,7 +1027,7 @@ module.exports = {
 				},
 			}).then(results =>
 				results.map(({ id, timestamp, type, account, currencyKey, side, amount, market, fee }) => ({
-					hash: id.split('-')[0],
+					hash: getHashFromId(id),
 					timestamp: Number(timestamp * 1000),
 					type,
 					account,
@@ -1142,7 +1133,7 @@ module.exports = {
 							hasPartialLiquidations,
 							collateralMinted,
 						}) => ({
-							id: Number(id.split('-')[0]),
+							id: Number(getHashFromId(id)),
 							account,
 							createdAt: new Date(Number(createdAt * 1000)),
 							closedAt: closedAt ? new Date(Number(closedAt * 1000)) : null,
@@ -1174,7 +1165,7 @@ module.exports = {
 				.then(results =>
 					results.map(({ account, liquidatedAmount, liquidator, liquidatedCollateral, loanId, id }) => ({
 						loanId: Number(loanId),
-						txHash: id.split('-')[0],
+						txHash: getHashFromId(id),
 						liquidatedCollateral: liquidatedCollateral / 1e18,
 						penaltyAmount: (liquidatedCollateral / 1e18) * 0.1,
 						liquidator,
@@ -1202,7 +1193,7 @@ module.exports = {
 				.then(results =>
 					results.map(({ account, liquidator, loanId, id, timestamp }) => ({
 						loanId: Number(loanId),
-						txHash: id.split('-')[0],
+						txHash: getHashFromId(id),
 						liquidator,
 						account,
 						timestamp: new Date(Number(timestamp * 1000)),
@@ -1314,7 +1305,7 @@ module.exports = {
 							destRoundIdAtPeriodEnd,
 							exchangeTimestamp,
 						}) => ({
-							hash: id.split('-')[0],
+							hash: getHashFromId(id),
 							from,
 							src: hexToAscii(src),
 							amount: amount / 1e18,
@@ -1423,7 +1414,7 @@ module.exports = {
 			}).then(results =>
 				results.map(({ id, time, account, amountLiquidated, snxRedeemed, liquidator }) => ({
 					id,
-					hash: id.split('-')[0],
+					hash: getHashFromId(id),
 					time: Number(time * 1000),
 					account,
 					liquidator,
